requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Cheminformatics
rdkit-pypi>=2023.3.1
PyTDC>=0.4.0

# Classical ML
xgboost>=2.0.0

# Deep Learning (Phase 3)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Demo / Deployment
gradio>=4.0.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
